requirements.txt
tensorflow
nvidia-cudnn-cu12
numpy
pandas
h5py
scipy
build
pytest
pdoc3
keras-multi-head
pydot
graphviz
backtrader
deap
pmdarima
protobuf
matplotlib
tensorflow-probability
keras.utils
tf-keras
tqdm
PyWavelets
flask
flask-cors
sqlalchemy
pyjwt
psutil
requests
